requirements.txt
packaging==24.2
wheel
discord.py
Pillow
